+++ ostap/fitting/roofit.py
@@ -326,15 +326,13 @@ def _rp_table_ ( plot , prefix = '' , title = '' ) :
         
         ## 
         row.append  ( name )
-
-
         
         row = tuple ( row  )
         
         table.append ( row )
 
     import ostap.logger.table as T
-    return T.table ( table , title = title, prefix = prefix , alignment = 'clccllll' )
+    return T.table ( table , title = title, prefix = prefix , alignment = 'clcclllw' )
 
 ROOT.RooPlot.table    =  _rp_table_
 ROOT.RooPlot.__str__  =  _rp_table_